Some minor style improvements
