+++ cypress.config.js
@@ -1,4 +1,4 @@
-const { defineConfig } = require("cypress");
+const { defineConfig } = require("cypress")
 
 module.exports = defineConfig({
   e2e: {
@@ -9,13 +9,11 @@ module.exports = defineConfig({
           console.log(message)
           return null
         },
-
         table(message) {
           console.table(message)
-
           return null
         }
       })
     },
   },
-});
+})
