+++ cypress/support/accessibility.js
@@ -1,6 +1,25 @@
 export const checkA11y = options => {
     cy.checkA11y(null, options, violations => {
-        console.log(`${violations.length} violation(s) detected`);
-        console.table(violations);
-    });
-};
+        console.log(`${violations.length} violation(s) detected`)
+        console.table(violations)
+    })
+}
+export function terminalLog(violations) {
+    cy.task(
+        'log',
+        `${violations.length} accessibility violation${
+            violations.length === 1 ? '' : 's'
+        } ${violations.length === 1 ? 'was' : 'were'} detected`
+    )
+    // pluck specific keys to keep the table readable
+    const violationData = violations.map(
+        ({ id, impact, description, nodes }) => ({
+            id,
+            impact,
+            description,
+            nodes: nodes.length
+        })
+    )
+    cy.task('table', violationData)
+}
+
